agentkit/pyproject.toml
[project]
name = "agentkit"
version = "0.3.0"
description = "A Python-native Agent framework with first-class Skill support and multi-LLM adapter"
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "README.md"
keywords = ["agent", "llm", "skill", "ai", "framework", "tool-use", "function-calling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.30.0"]
google = ["google-genai>=1.0.0"]
memory = ["mem0ai>=0.1.0"]
docker = ["docker>=7.0.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=1.0.0",
    "mem0ai>=0.1.0",
]

# CLI 入口：安装后可用 agentkit-docs 命令打开文档目录
[project.scripts]
agentkit-docs = "agentkit._cli:show_docs"

[project.urls]
Documentation = "https://github.com/agentkit/agentkit/tree/main/docs"
Repository = "https://github.com/agentkit/agentkit"

# ===== 构建配置 =====

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# 让 setuptools 将当前目录作为 agentkit 包
[tool.setuptools]
packages = [
    "agentkit",
    "agentkit.agents",
    "agentkit.llm",
    "agentkit.llm.adapters",
    "agentkit.memory",
    "agentkit.runner",
    "agentkit.safety",
    "agentkit.skills",
    "agentkit.tools",
    "agentkit.utils",
    "agentkit.examples",
    "agentkit.examples.ollama",
    "agentkit.examples.standard",
]

# 关键：告诉 setuptools 当前目录（.）包含 agentkit 包——即自身
[tool.setuptools.package-dir]
agentkit = "."

# 将 docs/ 和 examples/ 作为 package-data 一起打包
[tool.setuptools.package-data]
agentkit = [
    "docs/*.md",
    "examples/*.py",
    "examples/**/*.py",
    "examples/**/*.md",
]
